fix: resolve conflicts + rename file to user.ts
